Single/double precision scopes + integer ID standard
Replace per-call dtype arguments with ambient precision: new
precision()/data_dtype()/math_dtype() scope in _defaults.py, fed once
from RunConfig data_precision/math_precision ('single'/'double',
default single). Delete dtype_math and all cast_dtype params; mixture
classes derive X.dtype, retry ladder rebuilt to recompute in double.
Readers cast at load, assigner casts at fit, checkpoint refuses
cross-precision resume.

Standardize integer IDs (SIM_ID/LOCAL_IDX/GALAXY_ID/SNAP_ID/UNBOUND),
normalize boundness indices at set_boundness, guard the int64 map
narrowings, drop dead helpers. Validated: 556 unit + 20 integration
green, mock-sim single/double tracker identity empty.

+++ src/roadrunner/io/hdf5_particles.py
@@ -86,7 +86,7 @@ class HDF5ParticleWriter:
             ])
 
             scaler = StandardScaler()
-            scaled = scaler.fit_transform(coords.astype(np.float64, copy=False))
+            scaled = scaler.fit_transform(coords)
 
             scaler_grp = data.require_group("scaler")
             scaler_grp.create_dataset("mean", data=scaler.mean_)
